# Conflicts:
#	main.cpp
#	specificatie_1/VerkeersSituatieInlezen.h
#	src/elementen/Baan.cpp
#	src/elementen/Verkeerslicht.cpp
#	test/specificatie1/1_1_verkeerssituatie_inlezen/VerkeerssituatieInlezenTest.cpp

+++ src/elementen/Voertuiggenerator.cpp
@@ -4,6 +4,7 @@
 
 #include "Voertuiggenerator.h"
 
+
 std::string Voertuiggenerator::getBaan() const {
     return baan;
 }
@@ -12,10 +13,10 @@ void Voertuiggenerator::setBaan(const std::string &baan) {
     this->baan = baan;
 }
 
-int Voertuiggenerator::getFrequentie() const {
+double Voertuiggenerator::getFrequentie() const {
     return frequentie;
 }
 
-void Voertuiggenerator::setFrequentie(const int frequentie) {
+void Voertuiggenerator::setFrequentie(const double frequentie) {
     this->frequentie = frequentie;
 }
